+++ packages/keyring-api/src/xlm/types.test-d.ts
@@ -1,8 +1,8 @@
 import type { Extends } from '@metamask/keyring-utils';
 import { expectTrue } from '@metamask/keyring-utils';
 
-import type { XlmEoaAccount } from './types';
+import type { XlmAccount } from './types';
 import type { KeyringAccount } from '../api';
 
-// `XlmEoaAccount` extends `KeyringAccount`
-expectTrue<Extends<XlmEoaAccount, KeyringAccount>>();
+// `XlmAccount` extends `KeyringAccount`
+expectTrue<Extends<XlmAccount, KeyringAccount>>();