+++ packages/keyring-internal-api/src/types.ts
@@ -17,7 +17,7 @@ import {
   TrxAccountType,
   TrxEoaAccountStruct,
   XlmAccountType,
-  XlmEoaAccountStruct,
+  XlmAccountStruct,
 } from '@metamask/keyring-api';
 import { exactOptional, object } from '@metamask/keyring-utils';
 import type { Infer, Struct } from '@metamask/superstruct';
@@ -89,8 +89,8 @@ export const InternalTrxEoaAccountStruct = object({
   ...InternalAccountMetadataStruct.schema,
 });
 
-export const InternalXlmEoaAccountStruct = object({
-  ...XlmEoaAccountStruct.schema,
+export const InternalXlmAccountStruct = object({
+  ...XlmAccountStruct.schema,
   ...InternalAccountMetadataStruct.schema,
 });
 
@@ -116,7 +116,7 @@ export type InternalSolDataAccount = Infer<typeof InternalSolDataAccountStruct>;
 
 export type InternalTrxEoaAccount = Infer<typeof InternalTrxEoaAccountStruct>;
 
-export type InternalXlmEoaAccount = Infer<typeof InternalXlmEoaAccountStruct>;
+export type InternalXlmAccount = Infer<typeof InternalXlmAccountStruct>;
 
 export const InternalAccountStructs: Record<
   string,
@@ -128,7 +128,7 @@ export const InternalAccountStructs: Record<
   | Struct<InternalBtcP2trAccount>
   | Struct<InternalSolDataAccount>
   | Struct<InternalTrxEoaAccount>
-  | Struct<InternalXlmEoaAccount>
+  | Struct<InternalXlmAccount>
 > = {
   [`${EthAccountType.Eoa}`]: InternalEthEoaAccountStruct,
   [`${EthAccountType.Erc4337}`]: InternalEthErc4337AccountStruct,
@@ -138,7 +138,7 @@ export const InternalAccountStructs: Record<
   [`${BtcAccountType.P2tr}`]: InternalBtcP2trAccountStruct,
   [`${SolAccountType.DataAccount}`]: InternalSolDataAccountStruct,
   [`${TrxAccountType.Eoa}`]: InternalTrxEoaAccountStruct,
-  [`${XlmAccountType.Eoa}`]: InternalXlmEoaAccountStruct,
+  [`${XlmAccountType.Account}`]: InternalXlmAccountStruct,
 };
 
 export type InternalAccountTypes =
@@ -150,7 +150,7 @@ export type InternalAccountTypes =
   | InternalBtcP2trAccount
   | InternalSolDataAccount
   | InternalTrxEoaAccount
-  | InternalXlmEoaAccount;
+  | InternalXlmAccount;
 
 export const InternalAccountStruct = object({
   ...KeyringAccountStruct.schema,